+++ pkg/mem/set.amd64.go
@@ -16,16 +16,16 @@ func init() {
 // `n` bytes, with the value `v`.
 //
 //go:noescape
-func setU8UnalignedAVX2(a *byte, n uint64, v uint8)
+func setU8UnalignedAVX2(a *byte, n int, v uint8)
 
 // setU32UnalignedAVX2 sets every four bytes, starting at `a` and going on for
 // `n` bytes, with the value `v`.
 //
 //go:noescape
-func setU32UnalignedAVX2(a *byte, n uint64, v uint32)
+func setU32UnalignedAVX2(a *byte, n int, v uint32)
 
 // setU64UnalignedAVX2 sets every eight bytes, starting at `a` and going on for
 // `n` bytes, with the value `v`.
 //
 //go:noescape
-func setU64UnalignedAVX2(a *byte, n uint64, v uint64)
+func setU64UnalignedAVX2(a *byte, n int, v uint64)
